Move the messsages, requests and responses for each subsystem into submodules
